Remove invalid visited fragment optimization
When collecting fields in the query planner, if a fragment spread for
the same named fragment is encountered more than once, the occurrences
following the first one are completely ignored. This is inspired from
a similar optimization in the GraphQL execution algorithm (see
`visitedFragments` in https://spec.graphql.org/June2018/#sec-Field-Collection).

But while this is a valid optimization in GraphQL execution (it avoids
strictly redundant work), it is not in general for the query planner
as the collected fields include a scope, and that scope might be
different for different occurences of a spread.

For instance, with:
```graphql
{
  ... on X {
    ...S
  }
  ... on Y {
    ...S
  }
}
```
the two spread `...S` have different scopes (one has `X` as parent type,
the other has `Y`). Ignoring the 2nd occurences leads the resulting
query plan to completely (and incorrectly) omit the `... Y` condition.

This commit fixes the problem by always visiting all spreads when
collecting fields, thus removing the "visited fragments" tracking.

It would be possible to preserve the optimization in restricted cases
by taking the scope into account, and only ignore named fragments
spreads if we've already seen the same named fragments _with the same
scope_. It is however unclear if such optimization would ever
quick in for any "interesting" query. That is, such optimization
would kick in for something like:
```graphql
{
  ... on X {
    ...S
  }
  ... on X {
    ...S
  }
}
```
but it is obvious this query has some fat-fingered copy-paste "error".
And anything less obvious would likely end up with different scopes.

So at this point, we believe such optimization is simply not worth
the added compexity.