Remove the `enclosingScope` from scopes
It is currently unused, and it is currently a bit awkward to have it but
not include it in `identityKey`. It is also trivial to remove, so it
will be trivial to re-introduce if we start needing it. In the meantime,
it feels like it gets in the way more than anything.