Make `Schema.types/directives` return generators rather than maps
Since `Schema` exposes `Schema.type(name)`, `Schema.types` was only ever
use to iterate on the type values and the fact it was a map was more
getting in the way than anything. Having it return an iterable/generator
avoid the cruft of always calling `values()` on the return.

The patch also make those functions actual methods of `Schema` instead
of getters, as having getters returning generators feels weird. This
also also have an arg to those method to optionally include the
built-ins (even if the default is still not too).