Do not match full supergraph in tests
Doing so means any changes to the supergraph format always
trigger required changes to the tests, even if the changes
are completely unrelated to what is tested (say we bump a
version of a linked feature or similar).

Additionaly, by having the test only match the part of the
supergraph that are of interest to the test, it also makes
the test a bit more readable/focused.