Failing test
The test merges 2 versions of an input type where one of the field
is not merged, but also pass a default value for that input type
that include the non-merged field.

At the moment, this makes composition fails because the generated
supergraph will have the full default value, but it will be invalid
for the merged type. And the thrown error is kind of confusing.