WIP error paths from downstream errors
The `path` property on top-level errors is missing when the error originates in a subgraph fetch. This change attempts to recalculate the `path` by combining the "response path" with the downstream error path. It does some manual munging of the _entities field match the original operation and avoid exposing implementation details of the federated supergraph.