Refactor selection set implementation to be immutable
The current (before this commit) implementation of `SelectionSet` is
inherently mutable: its `SelectionSet.add` method mutate the selection
set it's called on. Which is at odds with the general query planner
algorithm, which is the main user of `SelectionSet`, but is
fundamentally based on dealing with immutable data (as we explore
various "path" options and the possible plans, most things are shared
(at least partly)).

This mismatch between the query planner that really want immutable
selection sets but an implementation that is mutable leads to
less than optimal code complexity and performance. More specificially,
if we want to use a `SelectionSet` into 2 separate branch of the
algorithm, we usually have to do defensive deep copies, which is
inefficient. And to mitigate those inefficiencies, we sometimes
don't do those copies, but this lead to code that fragile and easy
to break and has lead to bug previously (where we save a copy, but
then a branch mistakenly mutated and thus impacted another branch
mistakenly). A few tricks had been added to the implementation to
help mitigate the risks (the `freeze` behaviour, and the copy-on-write
in `FetchGroup`), but they add complexity and aren't always optimal
performance wise.

This commit thus rework the implementation/api of `SelectionSet` to make
it an immutable data structure. Doing so makes the code a lot safer
and easy to reason about. And it also provide some performance benefits:
compared to current `main` over a set of production schema/query, this
seem to provide around 20% improvement on average and up to almost 50%
improvment on some specific cases (disclaimer: those benchmark are
fairly unscientific so those numbers should be taken with a grain of
salt, but the numbers are clear enough that this is a net measurable
gain).