Optimise query plan generation when we cannot cross subgraphs
Often times, value types only references either "leaf" types or other
value types, but no entity type or root types. When a sub-part of a
query arrives to such a type, then we know that the rest of the
subselection is going to also be part of whichever fetch we're currently
building. We can use that knowledge to save work.

This is what this commit does. First, it pre-computes when building
the query planner which types have no reachable entity or root type
starting from them. Then, when computing query plans, it checks when
such types is reached, and when it is, it short-cuts the building of
the `GraphPath`s and `PathTree`s for the remaining sub-selections.

When subgraphs have a large number of value types (especially some
deeply nested ones), this can measurably speed up query plan
generation. This is particularly true when one federate either a single
subgraph or one subgraph is much large, which is a corner cases, but
can happen in the process of migrating an existin monolith to
federation).