Generalizes heurisitic to cut indirect paths to the root of paths
The planner has a number of heuristics whose goal is to cut the number
of concrete plans it has to fully generate and evaluate, which is
important to its performance. One of them cut paths that take
a "detour" in a subgraph when a more direct path exists, that is,
if we have some path:
```
T(S1) --[x]--> U(S1) --[y]--> V(S1)
```
then we ignore a path that, from `T` in S1 goes to some S2 subgraph to
get `x` to come back to S1, that is:
```
T(S1) --[key(id)] --> T(S2) --[x]--> U(S2) --[key(id)] --> U(S1) --[y]--> V(S1)
```
because the more direct option is always going to be better.

However, the code for this optimisation was not working for essentially
the same case but at the root of the path. That is, if we have path:
```
Query(S1) --[x]--> U(S1) --[y]--> V(S1)
```
then we were not ignoring:
```
Query(S2) --[x]--> U(S2) --[key(id)] --> U(S1) --[y]--> V(S1)
```
even though it similarly never going to be better (even if we must
get other fields after `x` in S2, it's still better to get 2 top-level
parallel fetch to both S1 and S2, rather than only getting to S2
to get to S1 next, as it's the same number of fetches but there is
more serialization).

Long story short, this commit slightly generalise the existing heuristic
so it cut this root-level case too.

Note that when a lot of queried field follow that pattern, this can
cut the number of plans considered drastically, so this optimisation
can also result in better generated plans due to the planner not falling
into its "panic" mode where it semi-arbitrarily cut options to avoid
exponential explosion, yielding potentially non-optimal plans in doing
so.

Fixes #2661.